+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [3.5.0] - 2025-03-29
+#### Added
+* Add `eslint-plugin-import`
+* Add `eslint-plugin-react-compiler`
+
+#### Changed
+* Expose base and react configs on their own entry points. Default entry point still merges both.
+
+#### Deprecated
+* *Nothing*
+
+#### Removed
+* *Nothing*
+
+#### Fixed
+* *Nothing*
+
+
 ## [3.4.1] - 2025-03-28
 #### Added
 * *Nothing*